+++ test/header_parser_tests/expected_bindings/_expected_comment_markup_bindings.dart
@@ -61,4 +61,8 @@ final class Com4 extends ffi.Struct {
   /// Single line field comment.
   @ffi.Float()
   external double b;
+
+  /// Comment on array member.
+  @ffi.Array.multi([3])
+  external ffi.Array<ffi.Int> c;
 }